# Full path expectations