--- a/composer.json
+++ b/composer.json
@@ -4,7 +4,7 @@
     "type": "library",
     "license": "MIT",
     "require": {
-        "laravel/framework": "^9.0",
+        "laravel/framework": "^9.36",
         "ext-dom": "*"
     },
     "require-dev": {
@@ -18,7 +18,8 @@
         }
     ],
     "conflict": {
-        "laravel/framework": "<9.36"
+        "laravel/framework": "<9.36.1",
+        "orchestra/testbench": "<7.11"
     },
     "autoload": {
         "psr-4": {