--- a/src/IconComponent.php
+++ b/src/IconComponent.php
@@ -5,10 +5,13 @@ declare(strict_types=1);
 namespace Orchid\Icons;
 
 use DOMDocument;
+use Illuminate\Container\Container;
+use Illuminate\Support\Arr;
 use Illuminate\View\Component;
 
 class IconComponent extends Component
 {
+
     /**
      * @var string|null
      */
@@ -40,21 +43,15 @@ class IconComponent extends Component
     public $id;
 
     /**
-     * @var IconFinder
-     */
-    private $finder;
-
-    /**
      * Icon tag
      *
      * @var string
      */
-    private $path;
+    public $path;
 
     /**
      * Create a new component instance.
      *
-     * @param IconFinder  $finder
      * @param string      $path
      * @param string|null $id
      * @param string|null $class
@@ -64,7 +61,6 @@ class IconComponent extends Component
      * @param string      $fill
      */
     public function __construct(
-        IconFinder $finder,
         string $path,
         string $id = null,
         string $class = null,
@@ -74,12 +70,11 @@ class IconComponent extends Component
         string $fill = 'currentColor'
     )
     {
-        $this->finder = $finder;
         $this->path = $path;
         $this->id = $id;
         $this->class = $class;
-        $this->width = $width ?? $finder->getDefaultWidth();
-        $this->height = $height ?? $finder->getDefaultHeight();
+        $this->width = $width;
+        $this->height = $height;
         $this->role = $role;
         $this->fill = $fill;
     }
@@ -87,15 +82,46 @@ class IconComponent extends Component
     /**
      * Get the view / contents that represent the component.
      *
-     * @return Icon
+     * @return string
      */
     public function render()
     {
-        $icon = $this->finder->loadFile($this->path);
+        $factory = $this->factory();
 
-        $content = $this->setAttributes($icon);
+        $factory->addNamespace(
+            '__components__orchid__icons',
+            $directory = Container::getInstance()['config']->get('view.compiled')
+        );
 
-        return new Icon($content);
+        $hash = sha1(collect($this->extractPublicProperties())->implode('-'));
+
+        if (! is_file($viewFile = $directory.'/'.$hash.'.blade.php')) {
+            if (! is_dir($directory)) {
+                mkdir($directory, 0755, true);
+            }
+
+            $contents = $this->renderIcon();
+
+            file_put_contents($viewFile, $contents);
+        }
+
+        return view('__components__orchid__icons::'.basename($viewFile, '.blade.php'));
+    }
+
+    /**
+     * @return string
+     * @throws \Illuminate\Contracts\Container\BindingResolutionException
+     */
+    private function renderIcon(): string
+    {
+        $finder = app()->make(IconFinder::class);
+
+        $this->width = $this->width ?? $finder->getDefaultWidth();
+        $this->height = $this->width ?? $finder->getDefaultHeight();
+
+        $icon = $finder->loadFile($this->path);
+
+        return $this->setAttributes($icon);
     }
 
     /**
@@ -113,16 +139,12 @@ class IconComponent extends Component
         $dom->loadXML($icon);
 
         /** @var \DOMElement $item */
-        $item = collect($dom->getElementsByTagName('svg'))->first();
+        $item = Arr::first($dom->getElementsByTagName('svg'));
 
-        collect($this->data())
-            ->except('attributes')
-            ->filter(function ($value) {
-                return $value !== null && is_string($value);
-            })
-            ->each(function (string $value, string $key) use ($item) {
-                $item->setAttribute($key, $value);
-            });
+        collect($this->extractPublicProperties())
+            ->except(['attributes'])
+            ->filter(static fn($value) => is_string($value))
+            ->each(static fn(string $value, string $key) => $item->setAttribute($key, $value));
 
         return $dom->saveHTML();
     }