… src/Icon.php
… tests/stubs/foo/regular/address-book.svg