Fixes https://github.com/orchidsoftware/platform/issues/2410

This PR is aimed at performance in finding a record of content about an icon.
To avoid going through each file and directory, the following changes have been made:


## Proposed Changes

  - Require that you specify the path for an embedded icon in the directory. For example: "prefix.directory.directory.icon"
  - Create real PHP files for `Blade`, in order to avoid going through the `DOMDocument` and setting size, color, etc. parameters.


## Backwards

This can affect applications that use strong nesting in icon directories. If used without nesting, a workaround to search the target directories has been added. 


## Benchmark

I measured on a 2018 Mac mini with an Intel processor and SSD drive on the Font Awesome dataset. The data is averaged and has been run several times.

```bash
Time of execution of one record:

// Master: 0.2111 sec. 
// PR:     0.1712 sec.


Time of execution of all records:

// Master:   12.765 sec.
// PR:       0.375  sec.
// PR:       0.702 sec. (At the first generation)
```